Entity Tree on his own file
Also, Entity Tree and Console, now shares some basic functionaly using a
base class called AbstractWindow (very lightweight)
